Move NullSocket
